fix admin

+++ src/src/Areas/Admin/Controllers/CategoriesController.cs
@@ -2,15 +2,18 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 using src.Models;
 using src.Repository;
 using src.Services;
 
 namespace src.Controllers
 {
+    [Area("Admin")]
     public class CategoriesController : Controller
     {
         private readonly ICategoryService _categoryService;
@@ -24,11 +27,25 @@ namespace src.Controllers
       
         public async Task<IActionResult> Index()
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             return View(await _categoryService.GetCategories());
         }
 
         public IActionResult Create()
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             return View();
         }
 
@@ -49,7 +66,13 @@ namespace src.Controllers
         // GET: Categories/Edit/5
         public async Task<IActionResult> Edit(int id)
         {
-
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
 
             var category = await _categoryService.GetCategoryById(id);
             if (category == null)
@@ -78,6 +101,13 @@ namespace src.Controllers
       
         public async Task<IActionResult> Delete(int id)
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             var categoryRepo = await _categoryService.DeleteCategory(id);
             if (categoryRepo == false) return NotFound();
             return RedirectToAction(nameof(Index));
