+++ src/src/Areas/Admin/Controllers/CitiesController.cs
@@ -2,15 +2,18 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 using src.Models;
 using src.Repository;
 using src.Services;
 
 namespace src.Controllers
 {
+    [Area("Admin")]
     public class CitiesController : Controller
     {
         private readonly ICityService _cityService;
@@ -25,12 +28,26 @@ namespace src.Controllers
        
         public async Task<IActionResult> Index()
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             return View(await _cityService.GetCities());
         }
 
         
         public async Task<IActionResult> Create()
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             ViewData["Regions"] = new SelectList(await _regionService.GetRegions(), "Id", "Name"); ;
             return View();
         }
@@ -52,7 +69,13 @@ namespace src.Controllers
       
         public async Task<IActionResult> Edit(int id)
         {
-
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
 
             var city = await _cityService.GetCityById(id);
             if (city == null)
@@ -83,7 +106,13 @@ namespace src.Controllers
       
         public async Task<IActionResult> Delete(int id)
         {
-
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             var c = await _cityService.DeleteCity(id);
             
             if (c == false) return NotFound();
