+++ src/src/Areas/Admin/Controllers/CountriesController.cs
@@ -2,15 +2,18 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 using src.Models;
 using src.Repository;
 using src.Services;
 
 namespace src.Controllers
 {
+    [Area("Admin")]
     public class CountriesController : Controller
     {
         private readonly ICountryService _countryService;
@@ -22,10 +25,24 @@ namespace src.Controllers
 
         public async Task<IActionResult> Index()
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             return View(await _countryService.GetCountries());
         }
         public async Task<IActionResult> Details(int id)
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             var country = await _countryService.GetCountryById(id);
             if (country == null)
             {
@@ -36,6 +53,13 @@ namespace src.Controllers
         }
         public IActionResult Create()
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             return View();
         }
 
@@ -52,6 +76,13 @@ namespace src.Controllers
         }  
         public async Task<IActionResult> Edit(int id)
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             var country = await _countryService.GetCountryById(id);
             if (country == null)
             {
@@ -80,6 +111,13 @@ namespace src.Controllers
         }
        public async Task<IActionResult> Delete(int id)
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             var deleteSuccess = await _countryService.DeleteCountry(id);
             if (!deleteSuccess)
             {
