+++ src/src/Areas/Admin/Controllers/RegionsController.cs
@@ -2,15 +2,18 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Newtonsoft.Json;
 using src.Models;
 using src.Repository;
 using src.Services;
 
 namespace src.Controllers
 {
+    [Area("Admin")]
     public class RegionsController : Controller
     {
         private readonly IRegionService _regionService;
@@ -23,12 +26,25 @@ namespace src.Controllers
         }
         public async Task<IActionResult> Index()
         {
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             return View(await _regionService.GetRegions());
         }
 
         public async Task<IActionResult> Details(int id)
         {
-          
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             var region = await _regionService.GetRegionById(id);
             if (region == null)
             {
@@ -39,7 +55,13 @@ namespace src.Controllers
         }
         public async Task<IActionResult> Create()
         {
-
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             IEnumerable<Country> cate = await _countryService.GetCountries();
             SelectList cateList = new SelectList(cate,"Id","Name");
 
@@ -63,7 +85,13 @@ namespace src.Controllers
         
         public async Task<IActionResult> Edit(int id)
         {
-
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
             var region = await _regionService.GetRegionById(id);
             if (region == null) return NotFound();
             ViewData["Country_id"] = new SelectList(await _countryService.GetCountries(), "Id", "Name", region.Country_id);
@@ -89,7 +117,13 @@ namespace src.Controllers
        
         public async Task<IActionResult> Delete(int id)
         {
-
+            var json = HttpContext.Session.GetString("user");
+            if (json == null)
+            {
+                return RedirectToAction("Index", "Login");
+            }
+            Models.Admin user = JsonConvert.DeserializeObject<Models.Admin>(json);
+            ViewBag.Username = user.Username;
 
             var region = await _regionService.DeleteRegion(id);
             if (region == false)
